added runtime console with commands, moved loader.LoadCmdArgs to misc.LoadCmdArgs, improved dev mode
